refactor(editor): refactored EditorService into maintainable chunks
