+++ honestcash-v2/src/app/modules/editor/interfaces/index.ts
@@ -1,0 +1,15 @@
+export interface DraftContext {
+  parentPostId?: number;
+  postId?: number;
+}
+
+export interface UploadImageResponse {
+  files: [{ url: string; }];
+}
+
+export interface UploadRemoteImageResponse {
+  success: number;
+  file: {
+    url: string;
+  };
+}
