+++ honestcash-v2/src/app/modules/editor/services/editor.service.ts
@@ -7,49 +7,9 @@ import Hashtag from '../../../shared/models/hashtag';
 import {HttpHeaders} from '@angular/common/http';
 import {ContentTypeFormDataHeader} from '../../../core/http/header.interceptor';
 import {LocalStorageToken} from '../../../core/helpers/localStorage';
-
-export interface DraftContext {
-  parentPostId?: number;
-  postId?: number;
-}
-
-export interface UploadImageResponse {
-  files: [{ url: string; }];
-}
-
-export interface UploadRemoteImageResponse {
-  success: number;
-  file: {
-    url: string;
-  };
-}
-
-export const API_ENDPOINTS = {
-  getPost: (id: number) => `/v2/post/${id}`,
-  getRelativePost: (id: number) => `/v2/post/relative/${id}`,
-  draft: (c: DraftContext = {}) =>
-    c.parentPostId ? `/v2/draft?parentPostId=${c.parentPostId}` : c.postId ? `/v2/post/${c.postId}` : '/v2/draft',
-  newDraft: () => `/v2/draft`,
-  savePostProperty: (p: Post, property: STORY_PROPERTIES) => `/v2/draft/${p.id}/${property}`,
-  saveDraft: (p: Post) => `/v2/draft/${p.id}/body`,
-  publishPost: (p: Post) => `/v2/draft/${p.id}/publish`,
-  uploadImage: () => `/upload/image`,
-  uploadRemoteImage: () => `/upload/image/remote`
-};
-
-export enum STORY_PROPERTIES {
-  Title = 'title',
-  Body = 'body',
-  BodyJSON = 'bodyJSON',
-  BodyAndTitle = 'bodyAndTitle',
-  Hashtags = 'hashtags',
-  PaidSection = 'paidSection',
-  HasPaidSection = 'hasPaidSection',
-  PaidSectionLinebreak = 'paidSectionLinebreak',
-  PaidSectionCost = 'paidSectionCost',
-}
-
-export const STORY_PREVIEW_KEY = 'HC_STORY_PREVIEW_BODY';
+import {API_ENDPOINTS} from '../shared/editor.endpoints';
+import {STORY_PREVIEW_KEY, STORY_PROPERTIES} from '../shared/editor.story-properties';
+import {DraftContext, UploadImageResponse, UploadRemoteImageResponse} from '../interfaces';
 
 @Injectable({
   providedIn: 'root'
@@ -61,23 +21,23 @@ export class EditorService {
   ) {
   }
 
-  getPost(id: number): Observable<Post> {
+  public getPost(id: number): Observable<Post> {
     return this.http.get<Post>(API_ENDPOINTS.getPost(id));
   }
 
-  getRelativePost(id: number) {
+  public getRelativePost(id: number) {
     return this.http.get<Post>(API_ENDPOINTS.getRelativePost(id));
   }
 
-  loadPostDraft(draftContext?: DraftContext): Observable<Post> {
+  public loadPostDraft(draftContext?: DraftContext): Observable<Post> {
     return this.http.get<Post>(API_ENDPOINTS.draft(draftContext));
   }
 
-  loadNewPostDraft(): Observable<Post> {
+  public loadNewPostDraft(): Observable<Post> {
     return this.http.post<Post>(API_ENDPOINTS.newDraft(), {});
   }
 
-  savePostProperty(post: Post, property: STORY_PROPERTIES): Observable<EmptyResponse> {
+  public savePostProperty(post: Post, property: STORY_PROPERTIES): Observable<EmptyResponse> {
     const body = {
       [property]: post[property]
     };
@@ -96,7 +56,7 @@ export class EditorService {
     return this.http.put<Post>(API_ENDPOINTS.savePostProperty(post, property), body);
   }
 
-  savePostPropertyLocally(property: STORY_PROPERTIES, value: any): void {
+  public savePostPropertyLocally(property: STORY_PROPERTIES, value: any): void {
     if (!value) {
       return;
     }
@@ -109,24 +69,24 @@ export class EditorService {
     this.localStorage.setItem(STORY_PREVIEW_KEY, JSON.stringify({property: value}));
   }
 
-  savePostLocally(story: Post) {
+  public savePostLocally(story: Post) {
     this.localStorage.setItem(STORY_PREVIEW_KEY, JSON.stringify(story));
   }
 
-  getLocallySavedPost(): Post {
+  public getLocallySavedPost(): Post {
     const data = JSON.parse(this.localStorage.getItem(STORY_PREVIEW_KEY));
     return data ? data : new Post();
   }
 
-  removeLocallySavedPost() {
+  public removeLocallySavedPost() {
     this.localStorage.removeItem(STORY_PREVIEW_KEY);
   }
 
-  publishPost(post: Post): Observable<Post> {
+  public publishPost(post: Post): Observable<Post> {
     return this.http.put<Post>(API_ENDPOINTS.publishPost(post), post);
   }
 
-  uploadImage(image: File): Observable<UploadImageResponse> {
+  public uploadImage(image: File): Observable<UploadImageResponse> {
     const formData = new FormData();
     formData.append('files[]', image, image.name);
 
@@ -137,8 +97,8 @@ export class EditorService {
     return this.http.post<UploadImageResponse>(API_ENDPOINTS.uploadImage(), formData, httpOptions);
   }
 
-  uploadRemoteImage(url: string) {
-    return this.http.post<UploadImageResponse>(API_ENDPOINTS.uploadRemoteImage(), {url});
+  public uploadRemoteImage(url: string): Observable<UploadRemoteImageResponse> {
+    return this.http.post<UploadRemoteImageResponse>(API_ENDPOINTS.uploadRemoteImage(), {url});
   }
 
   private transformTags(tags: Hashtag[]): string {
