+++ honestcash-v2/src/app/modules/editor/shared/editor.story-properties.ts
@@ -1,0 +1,13 @@
+export enum STORY_PROPERTIES {
+  Title = 'title',
+  Body = 'body',
+  BodyJSON = 'bodyJSON',
+  BodyAndTitle = 'bodyAndTitle',
+  Hashtags = 'hashtags',
+  PaidSection = 'paidSection',
+  HasPaidSection = 'hasPaidSection',
+  PaidSectionLinebreak = 'paidSectionLinebreak',
+  PaidSectionCost = 'paidSectionCost',
+}
+
+export const STORY_PREVIEW_KEY = 'HC_STORY_PREVIEW_BODY';
